allow building with packagekit-qt >= 1.0.0
Summary:
Currently fails to recognise packagekit-qt >= 1.0.0 with the message:
Could not find a configuration file for package "PackageKitQt5" that is compatible with requested version "0.9.5"
This results in the 'share' tab of Dolphin's folder properties simply
stating that samba is not installed, with the button which offers to
install this via packagekit not being present.
Currently known to affect at lease Kubuntu 18.04, 18.10, and Neon 18.04
packagekit-qt 0.9.5 dates from 2014, so regardless of more complicated
solutions, the version requirement can be dropped here.

BUG: 390599

Test Plan:
- Apply patch to Kubuntu CI builds of 18.08 branch.
- Apply patch to eariler releases for potential bugfix in Kubuntu
- Confirm that the button to install samba appears, installs samba,
  and that the samba sharing config options for the folder then appear.

Reviewers: cfeck, ngraham

Reviewed By: ngraham

Differential Revision: https://phabricator.kde.org/D15874