give add/remove errors a GUI
previously we'd ignore the returned errors leaving the user hanging when
adding a share fails and not even telling them why.

we now route errors on add/remove through somewhat generic
kmessagebox-based reporting tech.
unfortunately this is fairly excessive WRT string mapping because
ksambashare sports one enum for everything so API-wise a call to save()
could return UserShareAclOk but in practice that of course cannot happen
because that'd mean nothing. as such many of the strings are overly
generic because they cannot practically appear on weak context actions
like save().
very meh.

this depends on the new lastSystemErrorString() API in KIO to stringify
the majority of failure scenarios here, which come out of samba's 'net'
command

trying to share /tmp now produces a kmessagebox informing the user that
the dir cannot be shared

BUG: 334618
FIXED-IN: 20.12