port to qml and refactor to somewhat isolated pages
this is in preparation to add more "wizardy" behavior for scenarios
where samba isn't installed yet, such as adding the user to the samba
database.

the way the new ui works is that it's rendering a page stage, we push
pages on as they become relevant. longer term that'll be at least the
pages install->reboot->setpassword->acls

this also features some related improvements such as
- better const corectness for some member vars
- less ifdef samba_install
- the maximum share name is now capped at 60 characters (which is what
windows10 allows)
- there's a soft warning for the 60 character limit as well in the gui