repair share path handling
this fell over in the refactor so make it a bit more streamlined. the
context now gets a URL (file:///home/foo) and feeds that into the
resolveShare which constructs a path from it (/home/foo) to then feed to
ksambashare (which exclusively deals with local paths).
the path on the share instance is then the path to always use. to aid
with that we'll stop holding the url as a member

this unbreaks creating new shares which had no path set (this is now
done in resolveShare). also print that on applyChanges for easy
debugging