rejigger acl page and add a sheet to be shown when using a denial
denials always win so they shouldn't really be used unless one is
suuuuuuuuuuuuper sure about it.
to that end we'll throw up a warning that tries to explain that denials
probably aren't what the user wants and also that denying usually isn't
necessary because filesystem permissions still apply, so even when using
a liberal share ACL the filesystem ACL would still prevent other users
from reading one's private data for example

this was a bit awkward to implement because sheets need to overlay
something yet overlaying the columnlayout messes with the layout size
itself compacting the tableview away.
equally we cannot use a regular page because pages come with a whole
bunch of padding that looks very weird here and I've not found a way to
disable it either. instead we now have a new toplevel item which merely
acts as the thing the sheet can overlay. the item then contains the
previous layout and elements so they retain the sizing even when
overlayed

BUG: 422554
FIXED-IN: 20.12