add a group management page
to edit usershares the user needs to be member of the appropriate group.
detect when that is not the case (dir is not writable) and offer
automatic group management if at all viable (dir isn't group owned by
root and the group looks like something samba related).

to visualize this there's a new group page which is run after
installation and also on every start (it does a fairly cheap check;
never the less it is equipped with busy indication).

should the user not be able to write the usershare directory the user
can ask us to ask the kauth helper to make the necessary changes.

the entire backing logic is inside a groupmanager helper class that
evaluates the group status of the user

this also moves the reboot function back out of the samba installer into
the plugin. the group management also needs to reboot (or at least
re-log) to apply the changes so we need a more global way to issue
reboots

BUG: 407846
FIXED-IN: 20.12