Unify how the new Samba user is communicated to the user
The text talks about setting a password, but then various other things
talk about creating a user. Technically a user is indeed being created
behind the scenes, but this is something that the user doesn't really
need to know about. So let's just consistently communicate the
password-setting aspect to the user.

Also right-align the button to match other dialogs.