Load plugin only for directories
Specifying mime types in `X-KDE-ServiceTypes` no longer works in
KPropertiesDialog plugins. Instead the `MimeTypes` key in KPlugin
json metadata should be used (see frameworks/kio@909ab10d).
This commit introduces the direct use of json metadata and sets
`MimeTypes` to `inode/directory` so after this fix the plugin will
only be loaded for directories (as it should be).