reload combobox on reuse
tableview by default reuses delegates so we get state confused because
we don't bind our current index to the edit role (presumably because
we'd break the binding when the user selects something else anyway). in
any case because the delegates get reused we need to explicitly reload
our cached states when that happens. otherwise we have delegates
presenting the incorrect state (that is: out of sync with what is
actually in the model)

Fixes #2