Don't require that the samba shares group has "samba" in it
In principle this group could have any name at all. On some distros it's
"sambashare" but on others I'm familiar with, it's "usershare". Checking
to make sure that the group name isn't "root" is sensible, but beyond
that, we don't gain anything by being picky about this, and instead just
have the potential to add roadblocks to the initial setup. Let's not.


(cherry picked from commit ab7407cb9bcefe56afdd7ce65071e1601249e480)