Make initial setup much clearer
There are a lot of ways that initial Samba sharing setup can go wrong if
the distro hasn't set up everything perfectly out of the box. Right now
the wizard shows some error messages when problems are encountered, but
the logic fails to account for the full set of things that can go wrong
during initial setup, and as a result are sometimes displays misleading
error messages that can send the user down the wrong path.

This commit expands the set of conditions that are checked for, and
offers more verbose and actionable error messages specific to each of
them.

To accomplish this, the code is refactored to move nearly all of the
logic to the backend, with the QML side simply presenting the
information for whatever error was encountered.

The existing code already had logic for showing a friendly button to fix
the "you're not a member of the right group" error condition, which is
preserved here. TODOs are added in the place where other similar actions
to fix the other errors could be defined, which can be done later.

BUG: 425202
FIXED-IN: 22.12