Fix incorrect conditions and instructions in group setup
Right now the setup wizard checks for whether the group contains the
text "root", and if it does, it errors out and recommends that you
change the group owner to "usershares." But this isn't valid either; in
authhelper.cpp, it will only accept a group with the word "samba" in it,
plus a few other conditions.

This commit makes the frontend code for the wizard match the backend
conditions, so never recommends doing something invalid, which will then
fail.

BUG: 466786
FIXED-IN: 23.04