Remove wrong group check from group manager
Samba allows any group name for usershares. As long as the user is part
of the group the group name does not affect the function of usershares
at all.

The "net" man page even spells out "serverops" as example, which is
currently not accepted.

The remaining checks when creating the group is still bogus, but is
quite limited in scope, and has no effect when the group already exists.

BUG: 475946