If parent is mainWindow and rasterSurface, use Software rendering
When opening a properties dialog, or any dialog that uses
QtQuick (`QQuickView`), Dolphin will restart since it uses raster surface,
and QtQuick expects GPU accelerated surface.

During plugin construction, check if the parent window is using RasterSurface.
If so, set the QQuickWindow API to Software.

This fixes the restarting of Dolphin when opening properties, for example.

BUG:494627